+++ zarrs/src/array/codec/bytes_to_bytes.rs
@@ -6,6 +6,8 @@ pub mod blosc;
 pub mod bz2;
 #[cfg(feature = "crc32c")]
 pub mod crc32c;
+#[cfg(feature = "fletcher32")]
+pub mod fletcher32;
 #[cfg(feature = "gdeflate")]
 pub mod gdeflate;
 #[cfg(feature = "gzip")]
@@ -15,3 +17,6 @@ pub mod zstd;
 
 #[cfg(test)]
 pub mod test_unbounded;
+
+#[cfg(any(feature = "crc32c", feature = "fletcher32"))]
+mod strip_suffix_partial_decoder;